pyproject.toml
[project]
name = "deepvqe"
version = "0.1.0"
description = "DeepVQE: Joint AEC/NS/DR with soft delay estimation"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1.0",

    "einops>=0.7.0",
    "pyyaml>=6.0",
    "tensorboard>=2.15.0",
    "pesq>=0.0.4",
    "pystoi>=0.4.1",
    "scipy>=1.11.0",
    "soundfile>=0.12.0",
    "matplotlib>=3.8.0",
    "numpy>=1.24.0",
    "tqdm>=4.66.0",
    "gguf>=0.18.0",
]
